requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.27.0.post1
httpx==0.27.0
requests==2.31.0
python-dotenv==1.0.1